requirements.txt
psycopg2~=2.9.3
Pillow~=9.0.1
discord.py~=1.7.3